requirements.txt
streamlit==1.25.0
scikit-learn==1.2.2
nltk==3.8.1
PyPDF2==3.0.1
python-docx==0.8.11